Modified settings.sql

+++ pool/pumpScheduler/settings.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS `settings`;
 CREATE TABLE IF NOT EXISTS `settings` (
   `id` varchar(40) NOT NULL,
-  `value` tinyint(2) NOT NULL,
+  `value` smallint(6) NOT NULL,
   `userSetting` BOOLEAN NOT NULL ,
   `description` varchar(40) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
